algorithm now able to generate episodes and run an epoch on them

+++ training_algorithms/algorithm_deqn.py
@@ -57,7 +57,7 @@ class Algorithm_DEQN:
             s_t = s_tplus
 
         states_tensor = tf.stack(states)
-        return states_tensor
+        return tf.cast(states_tensor, dtype=tf.float32)
 
     def epoch(self, batches: tf.Tensor) -> tf.Tensor:
         """_summary_
